Move status polling while SEM alignment to the GUI
